+++ genesyscloud/telephony_providers_edges_edge_group/genesyscloud_telephony_providers_edges_edge_group_init.go
@@ -6,9 +6,7 @@ import (
 
 func SetRegistrar(l registrar.Registrar) {
 	l.RegisterDataSource("genesyscloud_telephony_providers_edges_edge_group", DataSourceEdgeGroup())
-
 	l.RegisterResource("genesyscloud_telephony_providers_edges_edge_group", ResourceEdgeGroup())
-
 	l.RegisterExporter("genesyscloud_telephony_providers_edges_edge_group", EdgeGroupExporter())
 
 }
