+++ genesyscloud/telephony_providers_edges_edge_group/genesyscloud_telephony_providers_edges_edge_group_init_test.go
@@ -23,13 +23,10 @@ func (r *registerTestInstance) registerTestResources() {
 
 	r.resourceMapMutex.Lock()
 	defer r.resourceMapMutex.Unlock()
-
 	providerResources["genesyscloud_telephony_providers_edges_edge_group"] = ResourceEdgeGroup()
-
-	// external package dependencies for outbound
+	// external package dependencies for Edges Edge group
 	providerResources["genesyscloud_telephony_providers_edges_site"] = edgeSite.ResourceSite()
 	providerResources["genesyscloud_telephony_providers_edges_trunkbasesettings"] = telephony.ResourceTrunkBaseSettings()
-
 	providerResources["genesyscloud_location"] = gcloud.ResourceLocation()
 
 }
@@ -38,9 +35,8 @@ func (r *registerTestInstance) registerTestDataSources() {
 
 	r.datasourceMapMutex.Lock()
 	defer r.datasourceMapMutex.Unlock()
-
 	providerDataSources["genesyscloud_telephony_providers_edges_edge_group"] = DataSourceEdgeGroup()
-	// external package dependencies for outbound
+	// external package dependencies for Edges Edge group
 	providerDataSources["genesyscloud_telephony_providers_edges_site"] = edgeSite.DataSourceSite()
 	providerDataSources["genesyscloud_telephony_providers_edges_trunkbasesettings"] = telephony.DataSourceTrunkBaseSettings()
 
@@ -49,19 +45,15 @@ func (r *registerTestInstance) registerTestDataSources() {
 func initTestresources() {
 	providerDataSources = make(map[string]*schema.Resource)
 	providerResources = make(map[string]*schema.Resource)
-
-	reg_instance := &registerTestInstance{}
-
-	reg_instance.registerTestDataSources()
-	reg_instance.registerTestResources()
+	regInstance := &registerTestInstance{}
+	regInstance.registerTestDataSources()
+	regInstance.registerTestResources()
 
 }
 
 func TestMain(m *testing.M) {
-
-	// Run setup function before starting the test suite for Outbound Package
+	// Run setup function before starting the test suite for Edges Edge group
 	initTestresources()
-
 	// Run the test suite for outbound
 	m.Run()
 }
