+++ genesyscloud/telephony_providers_edges_edge_group/resource_genesyscloud_telephony_providers_edges_edge_group.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"strings"
 	gcloud "terraform-provider-genesyscloud/genesyscloud"
 	"time"
 
@@ -299,4 +300,26 @@ func buildSdkTrunkBases(d *schema.ResourceData) *[]platformclientv2.Trunkbase {
 	}
 
 	return &returnValue
+}
+
+func GenerateEdgeGroupResourceWithCustomAttrs(
+	edgeGroupRes,
+	name,
+	description string,
+	managed,
+	hybrid bool,
+	otherAttrs ...string) string {
+	return fmt.Sprintf(`resource "genesyscloud_telephony_providers_edges_edge_group" "%s" {
+		name = "%s"
+		description = "%s"
+		managed = "%v"
+		hybrid = "%v"
+		%s
+	}
+	`, edgeGroupRes, name, description, managed, hybrid, strings.Join(otherAttrs, "\n"))
+}
+
+func GeneratePhoneTrunkBaseIds(userIDs ...string) string {
+	return fmt.Sprintf(`phone_trunk_base_ids = [%s]
+	`, strings.Join(userIDs, ","))
 }
